:recycle: refactor docker and kubectl main impl
